detect/http_stat_msg: move tests to tests/
