+++ src/detect-engine-mpm.c
@@ -52,7 +52,6 @@
 #include "detect-engine-payload.h"
 #include "detect-engine-hrhd.h"
 #include "detect-engine-hcd.h"
-#include "detect-engine-hsmd.h"
 #include "detect-engine-hscd.h"
 #include "detect-engine-hcbd.h"
 #include "detect-engine-hsbd.h"
