+++ src/detect-engine-register.c
@@ -157,7 +157,6 @@
 #include "detect-engine-hsbd.h"
 #include "detect-engine-hrhd.h"
 #include "detect-engine-hcd.h"
-#include "detect-engine-hsmd.h"
 #include "detect-engine-hscd.h"
 #include "detect-byte-extract.h"
 #include "detect-file-data.h"
